RecoverTags/requirements.txt
opencv-contrib-python==3.4.2.16   
opencv-python==4.0.0.21 
tqdm==4.31.1
numpy==1.16.3 
pandas==0.24.1 
Pillow==5.4.1
matplotlib==3.0.2